chore: configurar middleware CORS no backend

+++ backend/controllers/task.controller.js
@@ -35,7 +35,7 @@ export const updateTask = async (req, res) => {
     const task = req.body;
 
     if(!mongoose.Types.ObjectId.isValid(id)){
-        res.status(404).json({ success: false, message: "Invalid Task Id" });
+        return res.status(404).json({ success: false, message: "Invalid Task Id" });
     }
 
     try {
@@ -48,11 +48,16 @@ export const updateTask = async (req, res) => {
 
 export const deleteTask = async (req, res) => {
     const { id } = req.params;
+
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({ success: false, message: "Invalid Task Id" });
+    }
+
     try {
         await Task.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Task deleted" })
     } catch (error) {
         console.log("error in deleting tasks:", error.message);
-        res.status(404).json({ success: false, message: "Task not found" });
+        res.status(500).json({ success: false, message: "Server Error" });
     }
 }
